% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Some text}
\usage{
hello(x)
}
\arguments{
\item{x}{The name of the person to say hi to}
}
\value{
The output from \code{\link{print}}
}
\description{
Some other text
}
\details{
Prints "Hello, world!"
}
\examples{
hello('Foo')
\dontrun{hello('Bar')
}
}
